+++ lib/plug_http_cache.ex
@@ -331,11 +331,32 @@ defmodule PlugHTTPCache do
   def request(conn) do
     {
       conn.method,
-      Plug.Conn.request_url(conn),
+      request_url(conn),
       conn.req_headers,
       req_body(conn)
     }
   end
+
+  defp request_url(conn) do
+    IO.iodata_to_binary([
+      to_string(conn.scheme),
+      "://",
+      conn.host,
+      request_url_port(conn.scheme, conn.port),
+      conn.request_path,
+      request_url_qs(conn.query_string)
+    ])
+  end
+
+  defp request_url_port(:http, 80), do: ""
+  defp request_url_port(:https, 443), do: ""
+  defp request_url_port(_, port), do: [?:, Integer.to_string(port)]
+
+  # Conn's QS is by default not rfc3986 encoded…
+  defp request_url_qs(""), do: ""
+
+  defp request_url_qs(qs),
+    do: [??, qs |> :uri_string.dissect_query() |> :uri_string.compose_query()]
 
   defp response(conn) do
     {